==> 02_preprocessing_cn.ipynb <==
# import
import pandas as pd
import numpy as np
import re
import string
import pickle
from time import time

# functions
def extract_cn(text):
    pat = u'[\u4e00-\u9fa5\u3002\uff1b\uff0c\uff1a\u201c\u201d\uff08\uff09\u3001\uff1f\u300a\u300b]+'
    return ''.join(re.findall(pat, text))


# read table into dataframe
print("Loading...")
t0 = time()
path = '/Users/ouyangling/ds/metis/tradewar/01_data/chinese/'
df1 = pd.read_csv(path + 'corpus_cn_1.csv',sep = ';',encoding='utf8')
df2 = pd.read_csv(path + 'corpus_cn_2.csv',sep = ';',encoding='utf8')
df3 = pd.read_csv(path + 'corpus_cn_3.csv',sep = ';',encoding='utf8')

print("done in %0.3fs." % (time() - t0))


# concat multiple dfs
df = pd.concat([df1, df2, df3])

# cleaning
print("Cleaning data")
t0 = time()
df.drop_duplicates(inplace = True)
df = df.drop('geo', axis = 1)
df['text'] = df['text'].apply(extract_cn)
trade_war_documents_cn = df['text'].copy()
print("done in %0.3fs." % (time() - t0))

# pickle data
print("Pickling")
t0 = time()
with open('02_documents_cn.pickle', 'wb') as handle:
    pickle.dump(trade_war_documents_cn, handle, protocol=pickle.HIGHEST_PROTOCOL)
print("done in %0.3fs." % (time() - t0))


# pickle df
with open('02_df_cn.pickle', 'wb') as handle:
    pickle.dump(df, handle, protocol=pickle.HIGHEST_PROTOCOL)
