==> 02_preprocessing_en.ipynb <==
# import
import pandas as pd
import numpy as np
import re
import string
import pickle
from time import time

# functions
def remove_urls_mention_hashtag(s):
    # remove urls in string
    s1 = re.sub(r'(https?:\/\/)(\s)*(www\.)?(\s)*(\w+\.)*([\w\-\s]+\/)*([\w\-]+)\/?(\??\w+\s*=\w+\&?)*', '', s)
    s2 = re.sub(r'pic.twitter\S*', '', s1)
    s3 = re.sub(r'\@[a-zA-Z1-9]+', '', s2)
    #s3 = s2.replace('@', ' ')
    s4 = s3.replace('#', ' ')
    s5 = s4.replace('-', ' ')
    return s5


def remove_punc_digi_lower(text):
    # remove punctuation
    remove_punct = str.maketrans('', '', string.punctuation+'…'+'’'+'‘'+'“'+'”'+'–'+'—')
    text = text.translate(remove_punct)

    # remove digits and convert to lower case
    remove_digits = str.maketrans('', '', string.digits)
    text = text.lower().translate(remove_digits)
    
    return text

media_names = ['new york times','wall street journal','business insider',
               'washington post','chicago tribune','bloomberg','cnn',
               'cnn news','new york post','los angeles times','newsday']
def remove_media(text):
    for i in media_names:
        text = text.replace(i,' ')
    return text


# read table into dataframe
print("Loading...")
t0 = time()
path = '/Users/ouyangling/ds/metis/tradewar/01_data/english/'
df1 = pd.read_csv(path + 'corpus_en_1.csv',sep = ';')
df2 = pd.read_csv(path + 'corpus_en_2.csv',sep = ';')
df3 = pd.read_csv(path + 'corpus_en_3.csv',sep = ';')
df4 = pd.read_csv(path + 'corpus_en_4.csv',sep = ';')

print("done in %0.3fs." % (time() - t0))


# concat multiple dfs
df = pd.concat([df1, df2, df3, df4])

# pickle df
with open('02_df_en.pickle', 'wb') as handle:
    pickle.dump(df, handle, protocol=pickle.HIGHEST_PROTOCOL)

# cleaning
print("Cleaning data")
t0 = time()
df.drop_duplicates(inplace = True)
df = df.drop('geo', axis = 1)
df['text'] = df['text'].apply(str.strip)
df['text'] = df['text'].apply(remove_urls_mention_hashtag)
df['text'] = df['text'].apply(remove_punc_digi_lower)
df['text'] = df['text'].apply(remove_media)

trade_war_documents = df['text'].copy()
print("done in %0.3fs." % (time() - t0))

# pickle data
print("Pickling")
t0 = time()
with open('02_documents_en.pickle', 'wb') as handle:
    pickle.dump(trade_war_documents, handle, protocol=pickle.HIGHEST_PROTOCOL)
print("done in %0.3fs." % (time() - t0))
