==> 03_tokenize_cn.ipynb <==
# import
import pandas as pd
import numpy as np
import re
import string
import pickle
from time import time
from sklearn.utils import shuffle
import thulac
thu = thulac.thulac(filt=True)
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation

# functions
def tokenize_cn_join(text):
    # tokenize
    cut = thu.cut(text)
    #useful_tags = ['n','np','ns','ni','nz','v','j']
    tokens = [i[0] for i in cut]
    # remove stop words
    #stop_words = ['贸易战','是','要','有','能','来','贸易','贸易','中国','美国','战','么','特朗普','川普']
    #tokens_stop = [y for y in cut if y not in stop_words]

    return " ".join(tokens)

# read table into dataframe
print("Loading dataset...")
t0 = time()
with open('02_documents_cn.pickle', 'rb') as handle:
    trade_war_documents_cn = pickle.load(handle)
print("done in %0.3fs." % (time() - t0))

# tokenize
print("Tokenizing and joining...")
t0 = time()
trade_war_documents_cn = trade_war_documents_cn.apply(tokenize_cn_join)
print("done in %0.3fs." % (time() - t0))

# pickling
print("Pickling")
t0 = time()
with open('03_tokentext_cn.pickle', 'wb') as handle:
    pickle.dump(trade_war_documents_cn, handle, protocol=pickle.HIGHEST_PROTOCOL)
print("done in %0.3fs." % (time() - t0))
