==> 04_topics_en.ipynb <==
# import
import pandas as pd
import numpy as np
import re
import string
import pickle
import nltk
from time import time
from sklearn.utils import shuffle
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk.util import ngrams
from nltk.stem import SnowballStemmer, WordNetLemmatizer
from sklearn.model_selection import ParameterGrid
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation

%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()

from datetime import datetime
import datetime as dt

# read table into dataframe
print("Loading dataset...")
t0 = time()
with open('03_tokentext_en.pickle', 'rb') as handle:
    trade_war_tokentext = pickle.load(handle)
dataset = trade_war_tokentext
print("done in %0.3fs." % (time() - t0))

# parameters
n_features = 2000
n_components = 5
n_top_words = 10

# define stopwords
depun_gen_stop = [i.replace("'","") for i in stopwords.words('english')]
general_stopwords = list(set(stopwords.words('english') + depun_gen_stop))

web_stopwords = ['html','via','youtube','rt','twitter','tweet','tweets']
domain_stopwords = ['donald','trade','war','tradewar',
                    'realdonaldtrump','trump','trumps','dtj','djt','president','america','american']
media_stopwords = ['nyt','reuters','video','news','bloomberg','wsj','cnn','medium',
                   'newspaper','insider']
gabage_stopwords = ['g','doesnt','e','he','youre','dont','thats','could',
                    'really','would','may','much','many','everything','any',
                    'get','everyone','going','one','nobody','cant','nothing','lot','think','know',
                    'make','go','warhtml','say','day','week','look','said','want','plan','set','time',
                    'need','see','way','year','thing','people','move','state','take']

stop_words = general_stopwords + web_stopwords + domain_stopwords + media_stopwords + gabage_stopwords

# functions
def print_top_words(model, feature_names, n_top_words):
    for topic_idx, topic in enumerate(model.components_):
        message = "Topic #%d: " % topic_idx
        message += " ".join([feature_names[i]
                             for i in topic.argsort()[:-n_top_words - 1:-1]])
        print(message)
    print()

# tokenization and modeling

# Use tf-idf features for NMF.
print("Extracting tf-idf features for NMF...")
tfidf_vectorizer = TfidfVectorizer(max_df=0.4,
                                   max_features=n_features,
                                   stop_words=stop_words)
t0 = time()
tfidf = tfidf_vectorizer.fit_transform(dataset)
print("done in %0.3fs." % (time() - t0))
print('--------------------------------------')

# Fit the NMF model
print("Fitting the NMF model (Frobenius norm) with tf-idf features, "
      "n_components=%d..."
      % n_components)
t0 = time()
nmf = NMF(n_components=n_components, random_state=1,
          alpha=.1, l1_ratio=.5).fit(tfidf)
print("done in %0.3fs." % (time() - t0))

print("\nTopics in NMF model (Frobenius norm):")
tfidf_feature_names = tfidf_vectorizer.get_feature_names()
print_top_words(nmf, tfidf_feature_names, n_top_words)
print('--------------------------------------')

# get two metrices

doc_topic = nmf.transform(tfidf)
topic_word = nmf.components_

df_doc_topic = pd.DataFrame(doc_topic)
df_topic_word = pd.DataFrame(topic_word)

# load df

with open('02_df_en.pickle', 'rb') as handle:
    df = pickle.load(handle)
    
df = df.reset_index(drop=True)

# figure out each document's topic, filter out gabage

def assign_topic(indx):
    if doc_topic[indx].sum() == 0:
        return -1
    else:
        return np.argmax(doc_topic[indx])

# assign topics and filter out gabage

df['topic'] = [assign_topic(i) for i in df.index]
df = df[df['topic']!=-1]

# assign topic3 to topic1, topic4 to topic3

def change_topic(i):
    if i == 3:
        return 1
    else:
        return i

def change_topic2(i):
    if i == 4:
        return 3
    else:
        return i


df['topic'] = df['topic'].apply(change_topic)
df['topic'] = df['topic'].apply(change_topic2)

# map topic names
mapp = {0:'China & Fight',
        1:'Tariff & EU, Canada',
        2:'Stock Maket',
        3:'Farmer & Economy & Job'}

def map_name(num):
    return mapp[num]

df['topic_name'] = df['topic'].apply(map_name)

df.head()

# pickling

with open('04_df_en.pickle', 'wb') as handle:
    pickle.dump(df, handle, protocol=pickle.HIGHEST_PROTOCOL)

# pca plot

temp = df[['username','topic']]
temp = (temp.join(pd.DataFrame(tfidf.toarray()))
                        .drop('username',axis=1))

#y = temp['topic']
X = temp.drop('topic',axis=1)

from sklearn.decomposition import PCA
pca = PCA(n_components = 2)
PCAxy = pca.fit_transform(X)

# plt.figure(figsize = (16,12))
# plt.scatter(x[:,0], x[:,1], c=y)
# plt.show()

# plot data with seaborn

data = pd.DataFrame()
data['x']=PCAxy.T[0]
data['y']=PCAxy.T[1]
data['labels']=list(temp['topic'])

facet = sns.lmplot(data=data, x='x', y='y', hue='labels', 
                   fit_reg=False, legend=True, legend_out=True, size=12, aspect=1)

# get unique user names for location scraping

users = df['username'].unique()
with open('users.pickle', 'wb') as handle:
    pickle.dump(users, handle, protocol=pickle.HIGHEST_PROTOCOL)

# # get the most retweeted users top 1000

# top_re_users = (df.groupby('username',as_index=False).agg({'retweets':sum})
#                                 .sort_values(by='retweets',ascending=False)
#                                 .head(1000)['username'].tolist())
# # pickling

# with open('top_re_users.pickle', 'wb') as handle:
#     pickle.dump(top_re_users, handle, protocol=pickle.HIGHEST_PROTOCOL)
