==> 05_viz_pre_cn.ipynb <==
import pandas as pd
import pickle

from datetime import datetime
import datetime as dt

# load the documents df

with open('04_df_cn.pickle', 'rb') as handle:
    df = pickle.load(handle)

# tweet number

df.shape

df.head()

# unique users

df['username'].nunique()

# generate csv for tableau visualization

df['datetime'] = df['date'].apply(lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M"))
df.to_csv('05_viz_cn.csv')
